Image Queries and Service work communication (#8)
* query for menu image links
grabbing the images themself was a lot easier to do

* img query triggered on menu dom change
instead of having the button start the query, the menu images are queried prior on every menu change
or if the menu is gone, it clears the images

now the button just prints/alerts which images are currently grabbed
this will allow for us to use the button to initiate the analysis process and not worry abt images

* menu images queried go to pop up through background service worker
content script now sends chrome notification with all the found menu images currently on page
service worker recieves these messages and will later be used to send and recieve data from our api

right now the service worker automatically displays the currently grabbed images
in the new test page within the popup

things to worry about:
from console logs, it seems as if google chrome loads the first four or so images into the carousel first,
then it loads the rest into the carosel. should we somehow wait for all images? is there a way to identify
when all are loaded?

not so much worry, but the menu images are only sent when they are currently in view on google maps. do
we want to retain images so if somebody scrolls through a menu that doesnt contain the carousel, they
can still analyze?